esp32_security_sim/requirements.txt
# Core dependencies
flask==2.0.1
scapy==2.4.5
cryptography==36.0.0
numpy==1.21.4
requests==2.26.0

# Flask extensions
flask-wtf==1.0.0
flask-sqlalchemy==2.5.1

# Network analysis
netifaces==0.11.0

# Testing
pytest==6.2.5
coverage==6.1.2

# Security
pycrypto==2.6.1
